nathan student management/web_app/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
Pillow==10.0.1
pandas==2.0.3
matplotlib==3.7.2